refactor: keep SavedSearchAlerts under settings
